refine some interface for better expose api. (#52)

+++ include/LightGBM/meta.h
@@ -12,7 +12,7 @@ namespace LightGBM {
 /*! \brief Type of data size, it is better to use signed type*/
 typedef int32_t data_size_t;
 /*! \brief Type of score, and gradients */
-typedef double score_t;
+typedef float score_t;
 
 const score_t kMinScore = -std::numeric_limits<score_t>::infinity();
